+++ ext/agoo/gqleval.c
@@ -7,6 +7,7 @@
 #include "gqleval.h"
 #include "gqlintro.h"
 #include "gqljson.h"
+#include "gqlsub.h"
 #include "gqlvalue.h"
 #include "graphql.h"
 #include "http.h"
@@ -15,7 +16,9 @@
 #include "res.h"
 #include "sdl.h"
 #include "sectime.h"
+#include "sse.h"
 #include "text.h"
+#include "websocket.h"
 
 #define MAX_RESOLVE_ARGS	16
 
@@ -30,6 +33,7 @@ static const char	indent_str[] = "indent";
 static const char	json_content_type[] = "application/json";
 static const char	operation_name_str[] = "operationName";
 static const char	query_str[] = "query";
+static const char	subscription_str[] = "subscription";
 static const char	variables_str[] = "variables";
 
 
@@ -54,11 +58,14 @@ err_resp(agooRes res, agooErr err, int status) {
 		   code,
 		   at.year, at.mon, at.day, at.hour, at.min, at.sec, frac);
 
-    agoo_res_message_push(res, agoo_text_create(buf, cnt), true);
+    agoo_res_message_push(res, agoo_text_create(buf, cnt));
 }
 
+static char	ws_up[] = "HTTP/1.1 101 Switching Protocols\r\n";
+
 static void
-value_resp(agooRes res, gqlValue result, int status, int indent) {
+value_resp(agooReq req, gqlValue result, int status, int indent) {
+    agooRes		res = req->res;
     char		buf[256];
     struct _agooErr	err = AGOO_ERR_INIT;
     int			cnt;
@@ -69,6 +76,27 @@ value_resp(agooRes res, gqlValue result, int status, int indent) {
 	AGOO_ERR_MEM(&err, "response");
 	err_resp(res, &err, 500);
 	gql_value_destroy(result);
+	return;
+    }
+    if (NULL == result) {
+	switch (status) {
+	case 101:
+	    if (NULL == (text = agoo_text_append(text, ws_up, sizeof(ws_up) - 1)) ||
+		NULL == (text = agoo_ws_add_headers(req, text)) ||
+		NULL == (text = agoo_text_append(text, "\r\n", 2))) {
+
+		agoo_log_cat(&agoo_error_cat, "Failed to allocate memory for a response.");
+		return;
+	    }
+	    break;
+	case 200:
+	    text = agoo_sse_upgrade(req, text);
+	    break;
+	default:
+	    agoo_log_cat(&agoo_error_cat, "Did not expect an HTTP status of %d.", status);
+	    return;
+	}
+	agoo_res_message_push(res, text);
 	return;
     }
     if (AGOO_ERR_OK != gql_object_set(&err, msg, "data", result)) {
@@ -84,7 +112,7 @@ value_resp(agooRes res, gqlValue result, int status, int indent) {
     if (NULL == (text = agoo_text_prepend(text, buf, cnt))) {
 	agoo_log_cat(&agoo_error_cat, "Failed to allocate memory for a response.");
     }
-    agoo_res_message_push(res, text, true);
+    agoo_res_message_push(res, text);
 }
 
 gqlValue
@@ -361,13 +389,17 @@ gql_eval_get_hook(agooReq req) {
     gqlDoc		doc;
     gqlValue		result;
     gqlVar		vars = NULL;
+    gqlOpKind		default_kind = GQL_QUERY;
 
     if (NULL != (gq = agoo_req_query_value(req, indent_str, sizeof(indent_str) - 1, &qlen))) {
 	indent = (int)strtol(gq, NULL, 10);
     }
     if (NULL == (gq = agoo_req_query_value(req, query_str, sizeof(query_str) - 1, &qlen))) {
-	err_resp(req->res, &err, 500);
-	return;
+	if (NULL == (gq = agoo_req_query_value(req, subscription_str, sizeof(subscription_str) - 1, &qlen))) {
+	    err_resp(req->res, &err, 500);
+	    return;
+	}
+	default_kind = GQL_SUBSCRIPTION;
     }
     op_name = agoo_req_query_value(req, operation_name_str, sizeof(operation_name_str) - 1, &oplen);
     var_json = agoo_req_query_value(req, variables_str, sizeof(variables_str) - 1, &vlen);
@@ -380,7 +412,7 @@ gql_eval_get_hook(agooReq req) {
     }
     // Only call after extracting the variables as it terminates the string with a \0.
     qlen = agoo_req_query_decode((char*)gq, qlen);
-    if (NULL == (doc = sdl_parse_doc(&err, gq, qlen, vars))) {
+    if (NULL == (doc = sdl_parse_doc(&err, gq, qlen, vars, default_kind))) {
 	err_resp(req->res, &err, 500);
 	return;
     }
@@ -391,12 +423,44 @@ gql_eval_get_hook(agooReq req) {
     } else {
 	result = gql_doc_eval_func(&err, doc);
     }
-    gql_doc_destroy(doc);
     if (NULL == result) {
+	gql_doc_destroy(doc);
 	err_resp(req->res, &err, 500);
 	return;
     }
-    value_resp(req->res, result, 200, indent);
+    if (GQL_SUBSCRIPTION == doc->op->kind) {
+	int		status = 200;
+	gqlValue	sv;
+	const char	*subject;
+	gqlSub		sub;
+	gqlSel		query;
+
+	if (NULL == (sv = gql_object_get(result, "subject"))) {
+	    struct _agooErr	e;
+
+	    agoo_err_set(&e, AGOO_ERR_TYPE, "subscription did not return a subject");
+	    err_resp(req->res, &e, 400);
+	    gql_doc_destroy(doc);
+	    return;
+	}
+	subject = gql_string_get(sv);
+
+	if (AGOO_CON_WS == req->res->con_kind) {
+	    status = 101;
+	}
+	doc->op->kind = GQL_QUERY; // need so eval does the right thing
+	if (NULL == (sub = gql_sub_create(&err, req->res->con, subject, doc))) {
+	    err_resp(req->res, &err, 400);
+	    return;
+	}
+	agoo_server_add_gsub(sub);
+
+	value_resp(req, NULL, status, indent);
+
+	return;
+    }
+    gql_doc_destroy(doc);
+    value_resp(req, result, 200, indent);
 }
 
 static gqlValue
@@ -429,7 +493,7 @@ eval_post(agooErr err, agooReq req) {
 	return NULL;
     }
     if (0 == strncmp(graphql_content_type, s, sizeof(graphql_content_type) - 1)) {
-	if (NULL == (doc = sdl_parse_doc(err, req->body.start, req->body.len, vars))) {
+	if (NULL == (doc = sdl_parse_doc(err, req->body.start, req->body.len, vars, GQL_QUERY))) {
 	    return NULL;
 	}
     } else if (0 == strncmp(json_content_type, s, sizeof(json_content_type) - 1)) {
@@ -474,7 +538,7 @@ eval_post(agooErr err, agooReq req) {
 		    }
 		}
 	    }
-	    if (NULL == (doc = sdl_parse_doc(err, query, qlen, vars))) {
+	    if (NULL == (doc = sdl_parse_doc(err, query, qlen, vars, GQL_QUERY))) {
 		goto DONE;
 	    }
 	} else {
@@ -491,7 +555,9 @@ eval_post(agooErr err, agooReq req) {
     } else {
 	result = gql_doc_eval_func(err, doc);
     }
-
+    if (GQL_SUBSCRIPTION == doc->op->kind) {
+	result = NULL;
+    }
 DONE:
     gql_doc_destroy(doc);
     gql_value_destroy(j);
@@ -510,10 +576,12 @@ gql_eval_post_hook(agooReq req) {
     if (NULL != (s = agoo_req_query_value(req, indent_str, sizeof(indent_str) - 1, &len))) {
 	indent = (int)strtol(s, NULL, 10);
     }
-    if (NULL == (result = eval_post(&err, req))) {
+    if (NULL == (result = eval_post(&err, req)) && AGOO_ERR_OK != err.code) {
 	err_resp(req->res, &err, 400);
+    } else if (NULL == result) {
+	value_resp(req, result, 200, indent); // TBD return upgrade status if websocket, need to know from request what type it is
     } else {
-	value_resp(req->res, result, 200, indent);
+	value_resp(req, result, 200, indent);
     }
 }