+++ ext/agoo/websocket.c
@@ -28,7 +28,7 @@ agooText
 agoo_ws_add_headers(agooReq req, agooText t) {
     int		klen = 0;
     const char	*key;
-    
+
     t = agoo_text_append(t, up_con, sizeof(up_con) - 1);
     if (NULL != (key = agoo_con_header_value(req->header.start, req->header.len, "Sec-WebSocket-Key", &klen)) &&
 	klen + sizeof(ws_magic) < MAX_KEY_LEN) {
@@ -84,7 +84,7 @@ agoo_ws_decode(char *buf, size_t mlen) {
     bool	is_masked;
     uint8_t	*payload;
     uint64_t	plen;
-    
+
     b++; // op
     is_masked = 0x80 & *b;
     plen = 0x7F & *b;
@@ -105,7 +105,7 @@ agoo_ws_decode(char *buf, size_t mlen) {
     if (is_masked) {
 	uint8_t		mask[4];
 	uint64_t	i;
-	
+
 	for (i = 0; i < 4; i++, b++) {
 	    mask[i] = *b;
 	}
@@ -167,7 +167,7 @@ agoo_ws_calc_len(agooCon c, uint8_t *buf, size_t cnt) {
 bool
 agoo_ws_create_req(agooCon c, long mlen) {
     uint8_t	op = 0x0F & *c->buf;
-    
+
     if (NULL == (c->req = agoo_req_create(mlen))) {
 	agoo_log_cat(&agoo_error_cat, "Out of memory attempting to allocate request.");
 	return true;
@@ -200,7 +200,7 @@ void
 agoo_ws_req_close(agooCon c) {
     if (NULL != c->up && agoo_server.ctx_nil_value != c->up->ctx && c->up->on_close) {
 	agooReq	req = agoo_req_create(0);
-	    
+
 	req->up = c->up;
 	req->method = AGOO_ON_CLOSE;
 	req->hook = agoo_hook_create(AGOO_NONE, NULL, c->up->ctx, PUSH_HOOK, &agoo_server.eval_queue);
@@ -212,37 +212,27 @@ agoo_ws_req_close(agooCon c) {
 void
 agoo_ws_ping(agooCon c) {
     agooRes	res;
-    
+
     if (NULL == (res = agoo_res_create(c))) {
 	agoo_log_cat(&agoo_error_cat, "Memory allocation of response failed on connection %llu.", (unsigned long long)c->id);
     } else {
-	if (NULL == c->res_tail) {
-	    c->res_head = res;
-	} else {
-	    c->res_tail->next = res;
-	}
-	c->res_tail = res;
 	res->close = false;
 	res->con_kind = AGOO_CON_WS;
 	res->ping = true;
+	agoo_con_res_append(c, res);
     }
 }
 
 void
 agoo_ws_pong(agooCon c) {
     agooRes	res;
-    
+
     if (NULL == (res = agoo_res_create(c))) {
 	agoo_log_cat(&agoo_error_cat, "Memory allocation of response failed on connection %llu.", (unsigned long long)c->id);
     } else {
-	if (NULL == c->res_tail) {
-	    c->res_head = res;
-	} else {
-	    c->res_tail->next = res;
-	}
-	c->res_tail = res;
 	res->close = false;
 	res->con_kind = AGOO_CON_WS;
 	res->pong = true;
+	agoo_con_res_append(c, res);
     }
 }